rename InputModal dir name
